# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api Seletivo Seplag Gov-MT",
        "version": "0.1"
    },
    "paths": {
...
        "/api/musica": {
            "get": {
                "tags": [
                    "Musicas"
                ],
                "summary": "Todas as Musicas",
                "description": "Lista todas as Musicas",
                "operationId": "01de9509d3e6ccaa97ce184616faba9b",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Nº da página",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Musicas"
                ],
                "summary": "Cadastra nova Musica",
                "description": "Registra uma nova Musica",
                "operationId": "c703a59db1ae0961c25ba18b8721f70c",
                "parameters": [
                    {
                        "name": "mus_titulo",
                        "in": "query",
                        "description": "Titulo da Musica",
                        "required": true
                    },
                    {
                        "name": "album_id",
                        "in": "query",
                        "description": "Nº de identificação do Album",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Musica não encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/musica/{id}": {
            "get": {
                "tags": [
                    "Musicas"
                ],
                "summary": "Mostra uma Musica",
                "description": "Pesquisa por uma musica através do (id)",
                "operationId": "3d29bf3286bb4583067ce33fef6dec84",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Nº de identificação da musica",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Musica Encontrada",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Musica não encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Musicas"
                ],
                "summary": "Atualizar dados de uma Musica",
                "description": "Editar os dados de uma musica através do (id)",
                "operationId": "cbdaacd8ff8269d53b14a1b2f2eaf47f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da música a ser atualizada",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 5
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "album_id",
                                    "mus_titulo"
                                ],
                                "properties": {
                                    "album_id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "mus_titulo": {
                                        "type": "string",
                                        "example": "Nome musica"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "musica atualizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Musica atualizado com sucesso"
                                        },
                                        "musica": {
                                            "properties": {
                                                "mus_titulo": {
                                                    "type": "string",
                                                    "example": "Nome Musica"
                                                },
                                                "album_id": {
                                                    "type": "integer",
                                                    "example": "1"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Requisição inválida"
                    },
                    "404": {
                        "description": "Musica não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Musicas"
                ],
                "summary": "Exclui uma Musica",
                "description": "Exclui uma musica através do (id)",
                "operationId": "44a9d28fbf09b7140de33b217854e947",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do musica",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "musica excluída com sucesso",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Não foi possível excluir a musica",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/musica/{pesquisa}": {
            "get": {
                "tags": [
                    "Musicas"
                ],
                "summary": "Mostra uma Musica",
                "description": "Pesquisa por uma musica através do (pesquisa)",
                "operationId": "221ee3598723db33c96fa3dc1b0b8595",
                "parameters": [
                    {
                        "name": "pesquisa",
                        "in": "path",
                        "description": "Texto de pesquisa da musica",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "musica"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Musica Encontrada",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Musica não encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/regional": {
            "get": {
                "tags": [
                    "Regionais"
                ],
                "summary": "Todos as Unidades Regionais",
                "description": "Lista todos as Unidade Regionais",
                "operationId": "49685f8deed49afcb9948e7f3cdeb746",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Nº da página",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/regional/{pesquisa}": {
            "get": {
                "tags": [
                    "Regionais"
                ],
                "summary": "Pesquisa uma Regional pelo nome",
                "description": "Pesquisa por uma regional através do (pesquisa)",
                "operationId": "ac84b51a0e8dd6f65f9d24b8dde45eff",
                "parameters": [
                    {
                        "name": "pesquisa",
                        "in": "path",
                        "description": "Texto de pesquisa da regional",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "regional"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Regional Encontrada",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Musica não encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}